+++ ash/src/extensions/amdx/shader_enqueue.rs
@@ -25,7 +25,7 @@ impl crate::amdx::shader_enqueue::Device {
             pipeline_cache,
             create_infos.len() as u32,
             create_infos.as_ptr(),
-            allocation_callbacks.as_raw_ptr(),
+            allocation_callbacks.to_raw_ptr(),
             pipelines.as_mut_ptr(),
         );
         pipelines.set_len(create_infos.len());